Move `capitalizeFirstLetter` into `utilities/common`
