+++ src/utilities/common/capitalizeFirstLetter.ts
